Devops/chess/requirements.txt
python-chess>=1.999
Pillow>=10.2.0
opencv-python>=4.8.0
numpy>=1.26.0
cairosvg>=2.7.0
pyttsx3>=2.90
